Moved hod_components into occupation_models
