+++ halotools/empirical_models/test_empirical_models/__init__.py
@@ -2,6 +2,5 @@
 from __future__ import (absolute_import, division, print_function,
                         unicode_literals)
 
-from .test_hod_components import *
 from .test_mock_factories import *
 from .test_hod_factory import *
