+++ halotools/empirical_models/test_empirical_models/test_assembias.py
@@ -7,9 +7,8 @@ from copy import copy
 import numpy as np 
 from astropy.table import Table
 
-from ..hod_components import AssembiasZheng07Cens, AssembiasZheng07Sats, AssembiasLeauthaud11Cens, AssembiasLeauthaud11Sats
+from ..occupation_models.hod_components import *
 from .. import model_defaults
-from ..hod_components import Zheng07Cens, Leauthaud11Cens
 from ..sfr_models import BinaryGalpropInterpolModel
 
 from ...sim_manager import FakeSim
